Fix arithmetic overflow

+++ ARMDText/ARMDMessageParser.c
@@ -173,7 +173,7 @@ int ParseARMDMessage(ARMDMessageData* armd_data, ARMDHeaderInfo* armd_header_inf
 	{
 		GetValFromBuf(&armd_data->time, armd_parser_data, sizeof(DWORD)); //время
 		GetValFromBuf(&armd_data->events_len, armd_parser_data, sizeof(WORD)); //длина всех событий
-		if (armd_parser_data->index + armd_data->events_len < armd_parser_data->max_buf)
+		if ((size_t)armd_parser_data->index + armd_data->events_len < armd_parser_data->max_buf)
 		{
 			BYTE number_of_processes;
 			GetValFromBuf(&number_of_processes, armd_parser_data, sizeof(BYTE));//количество всех процессов
